requirements.txt
ariadne==0.11.0
black==19.10b0
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
graphql-core==3.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
MarkupSafe==1.1.1
# mysqlclient==1.4.6
PyMySQL==0.9.3
peewee==3.13.3
pyasn1==0.4.8
python-dotenv==0.13.0
python-jose==3.1.0
rsa==4.0
six==1.14.0
starlette==0.13.4
typing-extensions==3.7.4.2
Werkzeug==1.0.1